Added a standalone Accented keyboard layout bundle
I made the layout itself while working on my Bachelors degree in 2009, but never bothered to work on in further.
Adding it here just for the sake of it, but I do not know if the bundle will work out of the box, because I never
really bundled it until now and I dont have a Mac to test that it works fine.